@@ -9,14 +9,15 @@ class ReusableMiniCard extends StatelessWidget {
   final String tag;
   final Color cardColor;
 
-  ReusableMiniCard(
-      {@required this.icon,
-      @required this.cardColor,
-      @required this.tag,
-      @required this.title,
-      @required this.value,
-      @required this.titleColor,
-      @required this.iconColor});
+  ReusableMiniCard({
+    @required this.icon,
+    @required this.cardColor,
+    @required this.tag,
+    @required this.title,
+    @required this.value,
+    @required this.titleColor,
+    @required this.iconColor,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -24,15 +25,23 @@ class ReusableMiniCard extends StatelessWidget {
       tag: tag,
       child: Card(
         color: cardColor,
-        margin: EdgeInsets.symmetric(vertical: 9, horizontal: 8),
+        margin: EdgeInsets.symmetric(
+          vertical: 9,
+          horizontal: 8,
+        ),
         elevation: 16,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: <Widget>[
             Padding(
-              padding: const EdgeInsets.only(top: 15, left: 20),
+              padding: const EdgeInsets.only(
+                top: 15,
+                left: 20,
+              ),
               child: Icon(
                 icon,
                 color: iconColor,
@@ -40,7 +49,10 @@ class ReusableMiniCard extends StatelessWidget {
               ),
             ),
             Padding(
-              padding: const EdgeInsets.only(top: 15, left: 15),
+              padding: const EdgeInsets.only(
+                top: 15,
+                left: 15,
+              ),
               child: Text(
                 title.toUpperCase(),
                 style: TextStyle(
@@ -50,13 +62,19 @@ class ReusableMiniCard extends StatelessWidget {
               ),
             ),
             Padding(
-              padding: const EdgeInsets.only(top: 7, left: 15),
-              child: Text(
-                value.toString(),
-                style: TextStyle(
+              padding: const EdgeInsets.only(
+                top: 7,
+                left: 15,
+              ),
+              child: FittedBox(
+                child: Text(
+                  value.toString(),
+                  style: TextStyle(
                     fontSize: 50,
                     fontWeight: FontWeight.bold,
-                    color: Colors.blueGrey[800]),
+                    color: Colors.blueGrey[800],
+                  ),
+                ),
               ),
             ),
           ],
