@@ -1,33 +1,53 @@
 import 'package:flutter/material.dart';
 
-const backgroundColor = Color(0xFFECEFF1);
+const kBackgroundColor = Color(0xFFECEFF1);
 
 //Reusable card
-const reusableCardValueStyle = TextStyle(
+const kReusableCardValueStyle = TextStyle(
   fontSize: 30,
   fontWeight: FontWeight.bold,
 );
 
-const reusableCardTextStyle = TextStyle(fontSize: 23, color: Colors.white);
+const kReusableCardTextStyle = TextStyle(
+  fontSize: 23,
+  color: Colors.white,
+);
 
 //Rounded button
 
-const roundedButtonTextStyle =
-    TextStyle(color: Colors.white, fontSize: 17, fontWeight: FontWeight.bold);
+const kRoundedButtonTextStyle = TextStyle(
+  color: Colors.white,
+  fontSize: 17,
+  fontWeight: FontWeight.bold,
+);
 
 //main screens
 
-const mainTitleTextStyle =
-    TextStyle(fontSize: 30, color: Colors.white, fontWeight: FontWeight.bold);
+const kMainTitleTextStyle = TextStyle(
+  fontSize: 30,
+  color: Colors.white,
+  fontWeight: FontWeight.bold,
+);
 
-const lastUpdateTextStyle = TextStyle(
-    color: Color(0xFF546E7A),
-    fontSize: 15,
-    fontFamily: "OpenSans",
-    fontWeight: FontWeight.bold);
+const kLastUpdateTextStyle = TextStyle(
+  color: Color(0xFF546E7A),
+  fontSize: 15,
+  fontFamily: "OpenSans",
+  fontWeight: FontWeight.bold,
+);
 
-const cardTitleTextStyle = TextStyle(
-    color: Color(0xFF424242),
-    fontSize: 25,
-    fontWeight: FontWeight.bold,
-    fontFamily: "Muli");
+const kCardTitleTextStyle = TextStyle(
+  color: Color(0xFF424242),
+  fontSize: 25,
+  fontWeight: FontWeight.bold,
+  fontFamily: "Muli",
+);
+
+const String kApiKey =
+    "https://www.hpb.health.gov.lk/api/get-current-statistical";
+
+const String kSLimgLink =
+    "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRAHr6NvjNqDfxE2AdjX_0zwbVH2wfKP2ej9rcQFmB3_frsXuYhrboL-mMISD-YSoJ_xLbqirX39xb8xnc0Q3iQlEEqteAcugk&usqp=CAU&ec=45732300";
+
+const String kEarthFlagLink =
+    "https://upload.wikimedia.org/wikipedia/commons/thumb/e/ef/International_Flag_of_Planet_Earth.svg/1280px-International_Flag_of_Planet_Earth.svg.png";
