@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:trackingcorona/screens/loading_screen.dart';
-import 'package:trackingcorona/screens/sl_dashboard.dart';
-import 'package:trackingcorona/screens/global_dashboard.dart';
-import 'package:trackingcorona/screens/call_screen.dart';
-import 'package:trackingcorona/screens/faq_screen.dart';
+import './screens/loading_screen.dart';
+import './screens/sl_dashboard.dart';
+import './screens/global_dashboard.dart';
+import './screens/call_screen.dart';
+import './screens/faq_screen.dart';
 
 void main() => runApp(
       MaterialApp(
@@ -13,7 +13,7 @@ void main() => runApp(
           GlobalDashboard.id: (context) => GlobalDashboard(),
           LoadingScreen.id: (context) => LoadingScreen(),
           CallScreen.id: (context) => CallScreen(),
-          FAQScreen.id: (context) => FAQScreen()
+          FAQScreen.id: (context) => FAQScreen(),
         },
       ),
     );
