@@ -1,19 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:trackingcorona/components/reusable_card.dart';
-import 'package:trackingcorona/screens/global_dashboard.dart';
-import 'package:trackingcorona/components/custom_FAB.dart';
-import 'package:trackingcorona/components/under_Container.dart';
-import 'package:trackingcorona/components/title_container.dart';
-import 'package:trackingcorona/components/reusable_mini_card.dart';
-import 'package:trackingcorona/services/health_data.dart';
-import 'package:trackingcorona/components/constants.dart';
-import 'package:trackingcorona/components/main_drawer.dart';
+import './global_dashboard.dart';
+import '../widgets/reusable_card.dart';
+import '../widgets/floating_action_button.dart';
+import '../widgets/under_Container.dart';
+import '../widgets/title_container.dart';
+import '../widgets/reusable_mini_card.dart';
+import '../services/health_data.dart';
+import '../components/constants.dart';
+import '../widgets/main_drawer.dart';
 
 class SlDashboard extends StatefulWidget {
-  final Map dataMap;
-  SlDashboard({this.dataMap});
-
   static const String id = "SLDashboard";
   @override
   _SlDashboardState createState() => _SlDashboardState();
@@ -50,17 +47,22 @@ class _SlDashboardState extends State<SlDashboard> {
     _updateTime = data["updateTime"];
   }
 
-  @override
-  void initState() {
-    updateUI(widget.dataMap);
-    super.initState();
+  Future refreshUI() async {
+    HealthData healthData = HealthData();
+    var hlthData = await healthData.getHealthData();
+    setState(() {
+      updateUI(hlthData);
+    });
   }
 
   @override
   Widget build(BuildContext context) {
+    var dataMap = ModalRoute.of(context).settings.arguments;
+    updateUI(dataMap);
     return Scaffold(
       drawer: MainDrawer(),
-      floatingActionButton: CustomFAB(
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
+      floatingActionButton: FloatingActionBtn(
         label: "global",
         icon: FontAwesomeIcons.globeAmericas,
         color: Colors.blueGrey.shade900,
@@ -74,36 +76,34 @@ class _SlDashboardState extends State<SlDashboard> {
             "updateTime": _updateTime
           };
 
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) {
-                return GlobalDashboard(
-                  mapData: globalMap,
-                );
-              },
-            ),
-          );
+          Navigator.of(context)
+              .pushNamed(GlobalDashboard.id, arguments: globalMap);
         },
       ),
-      backgroundColor: backgroundColor,
+      backgroundColor: kBackgroundColor,
       body: SafeArea(
         child: RefreshIndicator(
-          onRefresh: () async {
-            HealthData healthData = HealthData();
-            var hlthData = await healthData.getHealthData();
-            setState(() {
-              updateUI(hlthData);
-            });
-          },
+          onRefresh: refreshUI,
           child: ListView(
             physics: const AlwaysScrollableScrollPhysics(),
             children: <Widget>[
-              TitleContainer(),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  TitleContainer(),
+                  Padding(
+                    padding: const EdgeInsets.only(right: 15),
+                    child: Image.network(
+                      kSLimgLink,
+                      height: 39,
+                    ),
+                  )
+                ],
+              ),
               Center(
                 child: Text(
                   "Last updated at $_updateTime",
-                  style: lastUpdateTextStyle,
+                  style: kLastUpdateTextStyle,
                 ),
               ),
               SizedBox(
@@ -118,7 +118,7 @@ class _SlDashboardState extends State<SlDashboard> {
                       padding: const EdgeInsets.only(top: 4),
                       child: Text(
                         "Total",
-                        style: cardTitleTextStyle,
+                        style: kCardTitleTextStyle,
                       ),
                     ),
                     ReusableCard(
@@ -180,7 +180,7 @@ class _SlDashboardState extends State<SlDashboard> {
                       padding: const EdgeInsets.only(top: 4),
                       child: Text(
                         "Daily",
-                        style: cardTitleTextStyle,
+                        style: kCardTitleTextStyle,
                       ),
                     ),
                     Row(
